feat: Add new pages and components
- Add UserPage component
- Add BookingPage component
- Add CustomersPage component
- Add RoomsPage component
- Add RoomTypesPage component
- Add RoomTypesGrid component
- Add NextUIProvider component
- Add createRoomType action
- Add NextUIProvider to RootLayout

This commit adds new pages and components to the project. The UserPage, BookingPage, CustomersPage, RoomsPage, and RoomTypesPage components are added to their respective directories. The RoomTypesGrid component is added to the RoomTypes directory. The NextUIProvider component is added to the providers directory. The createRoomType action is added to the actions file. Finally, the NextUIProvider component is added to the RootLayout component.